bulk_inference_pipeline/requirements-inference-bulk.txt
dill==0.3.4
flake8==4.0.1
joblib==1.2.0
numpy==1.23.4
pyflakes==2.4.0
python-dateutil==2.8.2
PyYAML==6.0
setuptools
spacy[cuda112]
spacy-transformers
google-auth==2.14.1
google.cloud-bigquery==3.3.6
google.cloud-storage==2.5.0
thinc_gpu_ops
GPUtil
--extra-index-url https://download.pytorch.org/whl/cu112
torch
